requirements.txt
mxnet-cu101mkl
mxboard
pydicom
boto3
pandas
tqdm
pillow